# fe/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "my-fam-tree API",
    "description": "Family platform API",
    "license": {
      "name": "BUSL-1.1",
      "identifier": "BUSL-1.1"
    },
    "version": "0.1.18"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
  "paths": {
...
    "/api/v1/families/{family_id}/transfer-owner": {
      "get": {
        "tags": [
          "owner-transfer"
        ],
        "operationId": "owner_transfer_status",
        "parameters": [
          {
            "name": "family_id",
            "in": "path",
            "description": "Family id (must match the active X-Family-Id)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending transfer or null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferStatusOptionalResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Path family_id does not match active family"
          },
          "403": {
            "description": "Insufficient role (admin / owner required)"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      },
      "post": {
        "tags": [
          "owner-transfer"
        ],
        "operationId": "owner_transfer_begin",
        "parameters": [
          {
            "name": "family_id",
            "in": "path",
            "description": "Family id (must match the active X-Family-Id)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeginReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transfer started; both confirmation emails dispatched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferStatusResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Path family_id does not match active family"
          },
          "403": {
            "description": "Owner role required"
          },
          "409": {
            "description": "A transfer is already pending"
          },
          "422": {
            "description": "Target is not a current admin in this family"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      },
      "delete": {
        "tags": [
          "owner-transfer"
        ],
        "operationId": "owner_transfer_cancel",
        "parameters": [
          {
            "name": "family_id",
            "in": "path",
            "description": "Family id (must match the active X-Family-Id)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending transfer cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Path family_id does not match active family"
          },
          "403": {
            "description": "Owner role required"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
    "/api/v1/families/{family_id}/transfer-owner/confirm": {
      "post": {
        "tags": [
          "owner-transfer"
        ],
        "summary": "Token-bearing: the supplied `token` IS the authorization. We require an\nauthenticated session (so the audit row has an `actor_user_id`) but do\nNOT bind to `X-Family-Id`, because the recipient may have a different\nactive family selected when they click the link from their email.",
        "operationId": "owner_transfer_confirm",
        "parameters": [
          {
            "name": "family_id",
            "in": "path",
            "description": "Family id (informational; the BE derives it from the token hash)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "One side confirmed; completes the transfer if both sides agreed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferStatusResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          },
          "404": {
            "description": "Token not found / already used"
          },
          "410": {
            "description": "Token expired"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BeginReq": {
        "type": "object",
        "required": [
          "to_user_id"
        ],
        "properties": {
          "to_user_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "ConfirmReq": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "NullResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {},
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
      "Pagination": {
        "type": "object",
        "description": "Cursor-based pagination block. `next_cursor` is `None` on the last page.",
        "required": [
          "limit",
          "returned"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "returned": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ResponseMeta": {
        "type": "object",
        "description": "Per-response metadata. Currently carries pagination cursors, the\n`request_id` for tracing-friendly client diagnostics, and any soft\nvalidation `warnings` attached by the handler.",
        "properties": {
          "pagination": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Pagination"
              }
            ]
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Warning"
            },
            "description": "Soft warnings. Serialized only when non-empty so the wire shape\nstays unchanged for handlers that don't produce any."
          }
        }
      },
...
      "TransferStatus": {
        "type": "object",
        "description": "Wire DTO for both single-transfer endpoints (`POST` begin / `POST` confirm\n/ `GET` status). `null` from `GET` is rendered as the outer `data: null`.",
        "required": [
          "id",
          "from_user_id",
          "to_user_id",
          "from_confirmed",
          "to_confirmed",
          "expires_at"
        ],
        "properties": {
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "from_confirmed": {
            "type": "boolean"
          },
          "from_user_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "to_confirmed": {
            "type": "boolean"
          },
          "to_user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "TransferStatusOptionalResponseBody": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TransferStatus"
              }
            ]
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
      "TransferStatusResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransferStatus"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
...
      "Warning": {
        "type": "object",
        "description": "Soft-validation warning.\n\nSurfaced under `meta.warnings` on the success envelope when the\ncaller's input violates a heuristic rule that we don't want to reject\noutright (e.g. parent-child age gap < 14 years). The `code` is a\nstable i18n key (`warning.<snake_case>`); `path` is an optional JSON\nPointer to the relevant field.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "cookie_access": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access"
      }
    }
  },
...
}